moved use_streams hook to a common place
