+++ src/platform/packages/shared/kbn-unified-chart-section-viewer/src/hooks/index.ts
@@ -9,3 +9,4 @@
 
 export { useEsqlQueryInfo } from './use_esql_query_info';
 export { useGridNavigation } from './use_grid_navigation';
+export { useStreamsNavigation } from './use_streams_navigation';
